production/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
cryptography==41.0.7
pyjwt==2.8.0
python-dotenv==1.0.0
bleach==6.1.0
